Extract execute_task to common module
