+++ cpcluster_client/Cargo.toml
@@ -12,4 +12,3 @@ serde_json = "1.0"
 env_logger = "0.10"
 log = "0.4"
 reqwest = { version = "0.11", features = ["rustls-tls"] }
-meval = "0.2"
