+++ cpcluster_common/Cargo.toml
@@ -7,3 +7,5 @@ edition = "2021"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tokio = { version = "1", features = ["full"] }
+reqwest = { version = "0.11", features = ["rustls-tls"] }
+meval = "0.2"
